Generate an arbitrary number of random samples from a normal distribution

Returns a data frame whose columns are random samples from a normal distribution

randomSample(mu = 0, sigma = 1, sampleSize = 25, numSamples = 1)

Arguments

mu
mean of normal distribution

sigma
standard deviation of normal distribution

sampleSize
size of each sample

numSamples
number of random samples to be generated

Value

A data frame whose columns are random samples from a normal distribution

Examples

head(randomSample(mu=1,sigma=4,sampleSize=100,numSamples=5))

#>           X1          X2          X3        X4          X5
#> 1  5.0797350 -4.26535058  2.03561001  1.383524  4.01901444
#> 2 -1.8083279  0.04581496  5.82987659  2.313460 -0.06933001
#> 3  4.8932631 -3.57251390 -0.07794545  5.864160  3.56832225
#> 4  0.6927294 -2.13790674 -1.79322715 -3.356316 -0.66930549
#> 5  4.5716997 -2.31107576 -4.95302377  3.177944 -3.86825650
#> 6 -2.1100124  4.77735022 10.85347650  2.845091  9.52945730